tiili: don't overwrite lit pixels with all-transparent ones
When shaper adjusted glyph rectangle it may overlap the previous
glyph. In these cases the pixel levels are to be added together.
It is probably enough to just skip writing pixel to memory when
the value of it is zero (all-transparent) - and that is done here
(that also was the original plan but I forgot).