with-labwc worktree build: use "cached" wlroots if available
Whenever new commits get added to upstream, the build make
creates new worktree. Every now and then the system wlroots
is not recent enough and one needs to use the one cloned in
subrojects -- which is empty in newly created git worktree.

If, after one clone, with-labwc/subprojects-cache/wlroots/
exists, and subprojects/wlroots does not exist in a worktree,
the former is symlinked as latter - before `meson setup`,
which - in its turn - (already) builds wlroots for the labwc
build that is done (a bit later) with `meson compile`.